+++ lib/super_good/solidus_taxjar.rb
@@ -16,6 +16,7 @@ module SuperGood
       attr_accessor :exception_handler
       attr_accessor :taxable_address_check
       attr_accessor :shipping_tax_label_maker
+      attr_accessor :line_item_tax_label_maker
     end
 
     self.discount_calculator = ::SuperGood::SolidusTaxJar::DiscountCalculator
@@ -25,5 +26,6 @@ module SuperGood
     }
     self.taxable_address_check = ->(address) { true }
     self.shipping_tax_label_maker = ->(shipment, shipping_tax) { "Sales Tax" }
+    self.line_item_tax_label_maker = ->(taxjar_line_item) { "Sales Tax" }
   end
 end